Updated stroke weight method

+++ lib/src/_core.dart
@@ -709,7 +709,7 @@ class Sketch {
   //------- End Shape/2D Primitives -----
 
   //----- Start Shape/Attributes ----
-  void strokeWeight(int newWeight) {
+  void strokeWeight(double newWeight) {
     if (newWeight < 0) {
       throw Exception('Stroke weight must be >= 0');
     }
